+++ webserver/Base64.h
@@ -8,7 +8,7 @@ std::string base64_encode_buf(unsigned char const* , unsigned int len);
 std::string base64_encode(std::string const& s);
 std::string base64_decode(std::string const& s);
 
-std::string base64url_encode_buf(unsigned char const* , unsigned int len);
+std::string base64url_encode_buf(unsigned char const* , size_t len);
 std::string base64url_encode(std::string const& s);
 std::string base64url_decode(std::string const& s);
